+++ views/pedidos/pedido-custo/_form-itens.php
@@ -100,7 +100,7 @@ use wbraganca\dynamicform\DynamicFormWidget;
                                                 items.each(function (index, elem) {
                                                     var priceValue = $(elem).find(".sumPart").val();
                                                     //Check if priceValue is numeric or something like that
-                                                    sum = parseInt(sum) + parseInt(priceValue);
+                                                    sum = parseFloat(sum) + parseFloat(priceValue);
                                                 });
                                                 //Assign the sum value to the field
                                                 $(".sum").val(sum);
@@ -150,7 +150,7 @@ $script = <<<EOD
         items.each(function (index, elem) {
             var priceValue = $(elem).find(".sumPart").val();
             //Check if priceValue is numeric or something like that
-            sum = parseInt(sum) + parseInt(priceValue);
+            sum = parseFloat(sum) + parseFloat(priceValue);
         });
         //Assign the sum value to the field
         $(".sum").val(sum);
